+++ src/main/java/com/mlt/api/apiproducts/repository/ProductImageRepository.java
@@ -3,5 +3,5 @@ package com.mlt.api.apiproducts.repository;
 import com.mlt.api.apiproducts.model.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
 }
